한 식물의 물주기 페이징 구현, 연관관계/엔티티그래프 삭제

+++ src/main/java/com/buckwheat/garden/data/dto/chemical/ChemicalDetail.java
@@ -8,5 +8,5 @@ import lombok.*;
 @AllArgsConstructor
 public class ChemicalDetail {
     private ChemicalDto chemical;
-    private long wateringSize;
+    private int wateringSize;
 }
